uploaded the script i made for my own historical and educational purposes
I quickly wrote a very rough and sloppy script (took about an hour, then another hour to work out the "bugs") - this semi-automates the entire process of modifying SystemUIGoogle.apk for all of the supported devices. From downloading them from GitRip, decompiling, modifying the package, recompiling, generating md5 hash, and copying all the resultant files into the proper folder structures, ready to be pushed to this repository. This will make it very easy and quick when the monthly updates come out.

I'm very aware I could have used loops and better logic, but like I said this was just a very quick script I wrote. Plus originally it was just for my own phone, then shortly after my wife's phone. Then eventually added support for the 3 and 3a series. In case GitRip or GitHub (or Goole) changes something, using loops and better logic would make it much easier to make future changes to the script. But oh well, it works for now. And this isn't a forever project anyway. But this was a fun little project to work on nonetheless. =)